[test] merge edge case tests into existing test files
