+++ tests/Handlers/BaseHandlerTest.php
@@ -200,4 +200,108 @@ class BaseHandlerTest extends BaseTestCase
             'arg2' => 'value2'
         ], $arguments);
     }
+
+    /**
+     * @return void
+     */
+    public function testIsPdfReturnsFalseForNonExistentFile(): void
+    {
+        $handler = new BaseHandler();
+        $this->assertFalse($handler->isPdf('/nonexistent/path/file.pdf'));
+    }
+
+    /**
+     * @return void
+     */
+    public function testIsPdfReturnsFalseForEmptyString(): void
+    {
+        $handler = new BaseHandler();
+        $this->assertFalse($handler->isPdf(''));
+    }
+
+    /**
+     * @return void
+     */
+    public function testIsPdfReturnsFalseForDirectoryPath(): void
+    {
+        $handler = new BaseHandler();
+        $this->assertFalse($handler->isPdf(sys_get_temp_dir()));
+    }
+
+    /**
+     * @return void
+     */
+    public function testGetTmpFileContainsTmpPath(): void
+    {
+        $handler = new BaseHandler();
+        $tmpFile = $handler->getTmpFile();
+        $this->assertStringStartsWith($handler->getTmpPath(), $tmpFile);
+    }
+
+    /**
+     * @return void
+     */
+    public function testGetTmpFileWithCustomFilenameContainsPrefix(): void
+    {
+        $handler = new BaseHandler();
+        $tmpFile = $handler->getTmpFile('custom');
+        $this->assertStringContainsString('ghostscript_tmp_file_custom', $tmpFile);
+        $this->assertStringEndsWith('.pdf', $tmpFile);
+    }
+
+    /**
+     * @return void
+     */
+    public function testForceClearDeletesTmpPrefixedFiles(): void
+    {
+        $handler = new BaseHandler();
+        $tmpFile = tempnam($handler->getTmpPath(), BaseHandler::TMP_FILE_PREFIX);
+        @rename($tmpFile, $tmpFile . '.pdf');
+        $tmpFile .= '.pdf';
+        $handler->clearTmpFiles(true);
+        $this->assertFalse(file_exists($tmpFile));
+    }
+
+    /**
+     * @return void
+     */
+    public function testOptionsToCommandWithKeyValuePairs(): void
+    {
+        $handler = new BaseHandler();
+        $handler->setOptions(['-dCompatibilityLevel' => '1.4']);
+        $result = $handler->optionsToCommand('gs');
+        $this->assertEquals('gs -dCompatibilityLevel=1.4', $result);
+    }
+
+    /**
+     * @return void
+     */
+    public function testOptionsToCommandWithNumericKeys(): void
+    {
+        $handler = new BaseHandler();
+        $handler->setOptions(['-dSAFER', '-dBATCH']);
+        $result = $handler->optionsToCommand('gs');
+        $this->assertEquals('gs -dSAFER -dBATCH', $result);
+    }
+
+    /**
+     * @return void
+     */
+    public function testOptionsToCommandWithEmptyOptionsReturnsOriginal(): void
+    {
+        $handler = new BaseHandler();
+        $this->assertEquals('gs -sDEVICE=pdfwrite', $handler->optionsToCommand('gs -sDEVICE=pdfwrite'));
+    }
+
+    /**
+     * @return void
+     */
+    public function testValidateBinPathThrowsForNonExistentPath(): void
+    {
+        $this->expectException(InvalidException::class);
+        $this->expectExceptionCode(InvalidException::CODE_FILEPATH);
+        $handler = new BaseHandler();
+        $handler->setBinPath('/nonexistent/gs');
+        $handler->validateBinPath();
+    }
 }
