+++ tests/Handlers/MergeHandlerTest.php
@@ -124,4 +124,46 @@ class MergeHandlerTest extends BaseTestCase
             dirname(__DIR__, 2) . '/files/merge/part_3.pdf'
         ]);
     }
+
+    /**
+     * @return void
+     */
+    public function testExecuteWithEmptyFilesArrayThrowsHandlerException(): void
+    {
+        $this->expectException(HandlerException::class);
+        $this->expectExceptionCode(HandlerException::CODE_EXECUTE);
+        $handler = new MergeHandler();
+        $handler->setBinPath($this->getEnv('GS_BIN_PATH'));
+        $handler->execute(dirname(__DIR__, 2) . '/files/merge', 'res.pdf', []);
+    }
+
+    /**
+     * @return void
+     */
+    public function testExecuteWithAllInvalidFilesThrowsHandlerException(): void
+    {
+        $this->expectException(HandlerException::class);
+        $this->expectExceptionCode(HandlerException::CODE_EXECUTE);
+        $handler = new MergeHandler();
+        $handler->setBinPath($this->getEnv('GS_BIN_PATH'));
+        $handler->execute(dirname(__DIR__, 2) . '/files/merge', 'res.pdf', [
+            dirname(__DIR__, 2) . '/files/merge/nonexistent1.pdf',
+            dirname(__DIR__, 2) . '/files/merge/nonexistent2.pdf',
+        ]);
+    }
+
+    /**
+     * @return void
+     */
+    public function testExecuteWithOnlyOneValidFileMergesSuccessfully(): void
+    {
+        $path = dirname(__DIR__, 2) . '/files/merge';
+        $handler = new MergeHandler();
+        $handler->setBinPath($this->getEnv('GS_BIN_PATH'));
+        $result = $handler->execute($path, 'single.pdf', [
+            dirname(__DIR__, 2) . '/files/merge/part_1.pdf',
+            dirname(__DIR__, 2) . '/files/merge/nonexistent.pdf',
+        ]);
+        $this->assertFileExists($result);
+    }
 }
